Create Space-time Folds

Create spatial, temporal or spatio-temporal Folds for cross validation

CreateSpacetimeFolds(
  x,
  spacevar = NA,
  timevar = NA,
  k = 10,
  class = NA,
  seed = sample(1:1000, 1)
)

Arguments

x

data.frame containing spatio-temporal data

spacevar

Character indicating which column of x identifies the spatial units (e.g. ID of weather stations)

timevar

Character indicating which column of x identifies the temporal units (e.g. the day of the year)

k

numeric. Number of folds. If spacevar or timevar is NA and a leave one location out or leave one time step out cv should be performed, set k to the number of unique spatial or temporal units.

class

Character indicating which column of x identifies a class unit (e.g. land cover)

seed

numeric. See ?seed

Value

A list that contains a list for model training and a list for model validation that can directly be used as "index" and "indexOut" in caret's trainControl function

Details

Using "class" is helpful in the case that data are clustered in space and are categorical. E.g This is the case for land cover classifications when training data come as training polygons. In this case the data should be split in a way that entire polygons are held back (spacevar="polygonID") but at the same time the distribution of classes should be similar in each fold (class="LUC").

Note

Standard k-fold cross-validation can lead to considerable misinterpretation in spatial-temporal modelling tasks. This function can be used to prepare a Leave-Location-Out, Leave-Time-Out or Leave-Location-and-Time-Out cross-validation as target-oriented validation strategies for spatial-temporal prediction tasks. See Meyer et al. (2018) for further information.

References

Meyer, H., Reudenbach, C., Hengl, T., Katurji, M., Nauß, T. (2018): Improving performance of spatio-temporal machine learning models using forward feature selection and target-oriented validation. Environmental Modelling & Software 101: 1-9.

See also

Author

Hanna Meyer

Examples

dat <- get(load(system.file("extdata","Cookfarm.RData",package="CAST")))
### Prepare for 10-fold Leave-Location-and-Time-Out cross validation
indices <- CreateSpacetimeFolds(dat,"SOURCEID","Date")
str(indices)
#> List of 2
#>  $ index   :List of 10
#>   ..$ : int [1:147368] 1 2 3 4 5 11 12 13 14 15 ...
#>   ..$ : int [1:150398] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:153974] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:149032] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:151376] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:144930] 6 7 8 9 10 11 12 13 14 15 ...
#>   ..$ : int [1:150515] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:146991] 291 294 296 298 300 302 303 305 306 308 ...
#>   ..$ : int [1:144393] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:145016] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ indexOut:List of 10
#>   ..$ : int [1:1971] 1275 1284 1293 1302 1310 1406 1414 1421 1429 1437 ...
#>   ..$ : int [1:1665] 1135 1136 1144 1145 1153 1154 1162 1168 1224 1225 ...
#>   ..$ : int [1:1265] 2551 2561 2571 2581 2590 2601 2610 2621 2631 2642 ...
#>   ..$ : int [1:1729] 1678 1684 1696 1703 1771 1777 1789 1796 4873 4876 ...
#>   ..$ : int [1:1474] 1320 1329 1338 1342 1347 1350 1355 1364 1373 1381 ...
#>   ..$ : int [1:2191] 508 512 516 520 524 577 581 585 589 593 ...
#>   ..$ : int [1:1565] 2183 2191 2201 2211 2239 2254 2269 2285 2315 2330 ...
#>   ..$ : int [1:2005] 36 37 38 39 40 56 57 58 59 60 ...
#>   ..$ : int [1:2240] 386 389 392 395 398 815 819 823 827 831 ...
#>   ..$ : int [1:2208] 315 318 321 324 327 417 420 423 426 429 ...
### Prepare for 10-fold Leave-Location-Out cross validation
indices <- CreateSpacetimeFolds(dat,spacevar="SOURCEID")
str(indices)
#> List of 2
#>  $ index   :List of 10
#>   ..$ : int [1:166676] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:164875] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:170172] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:163750] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:161641] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:167860] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:160317] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:160848] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:165513] 291 294 296 298 300 302 303 305 306 308 ...
#>   ..$ : int [1:167238] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ indexOut:List of 10
#>   ..$ : int [1:16534] 1060 1068 1076 1083 1089 1097 1105 1113 1120 1126 ...
#>   ..$ : int [1:18335] 526 530 532 536 538 542 544 548 550 554 ...
#>   ..$ : int [1:13038] 303 306 309 312 315 318 321 324 327 330 ...
#>   ..$ : int [1:19460] 291 294 296 298 300 302 305 308 311 314 ...
#>   ..$ : int [1:21569] 1205 1213 1253 1261 1298 1306 1342 1350 1385 1393 ...
#>   ..$ : int [1:15350] 2207 2277 2354 2490 2589 2673 2696 2720 2759 2779 ...
#>   ..$ : int [1:22893] 1138 1147 1156 1164 1170 1180 1190 1200 1210 1218 ...
#>   ..$ : int [1:22362] 1179 1189 1199 1209 1217 1227 1237 1247 1257 1265 ...
#>   ..$ : int [1:17697] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:15972] 1061 1069 1077 1084 1090 1098 1106 1114 1121 1127 ...
### Prepare for leave-One-Location-Out cross validation
indices <- CreateSpacetimeFolds(dat,spacevar="SOURCEID",
k=length(unique(dat$SOURCEID)))
str(indices)
#> List of 2
#>  $ index   :List of 42
#>   ..$ : int [1:177689] 291 294 296 298 300 302 303 305 306 308 ...
#>   ..$ : int [1:178560] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178144] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178043] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178096] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178715] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178887] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:180331] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178005] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178619] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:179049] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178599] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178969] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178547] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178108] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178489] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178629] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178747] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178004] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:177939] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178483] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178609] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178284] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178620] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:179932] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178231] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:179250] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178643] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178506] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:179638] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178773] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178164] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:179085] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:179431] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:178994] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:179373] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:180191] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:179380] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:179715] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:179701] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:180629] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:179809] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ indexOut:List of 42
#>   ..$ : int [1:5521] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ : int [1:4650] 291 294 296 298 300 302 305 308 311 314 ...
#>   ..$ : int [1:5066] 303 306 309 312 315 318 321 324 327 330 ...
#>   ..$ : int [1:5167] 448 452 456 460 464 468 472 476 480 484 ...
#>   ..$ : int [1:5114] 526 532 538 544 550 1062 1070 1078 1085 1091 ...
#>   ..$ : int [1:4495] 530 536 542 548 554 558 562 566 570 574 ...
#>   ..$ : int [1:4323] 1060 1068 1076 1083 1089 1097 1105 1113 1120 1126 ...
#>   ..$ : int [1:2879] 1061 1069 1077 1084 1090 1098 1106 1114 1121 1127 ...
#>   ..$ : int [1:5205] 1138 1147 1156 1164 1170 1180 1190 1200 1210 1218 ...
#>   ..$ : int [1:4591] 1179 1189 1199 1209 1217 1227 1237 1247 1257 1265 ...
#>   ..$ : int [1:4161] 1205 1213 1253 1261 1298 1306 1342 1350 1385 1393 ...
#>   ..$ : int [1:4611] 2172 2221 2297 2374 2389 2403 2419 2435 2452 2472 ...
#>   ..$ : int [1:4241] 2183 2191 2201 2211 2239 2254 2269 2285 2315 2330 ...
#>   ..$ : int [1:4663] 2207 2277 2354 2759 2779 2802 2826 2848 2871 2891 ...
#>   ..$ : int [1:5102] 2218 2233 2247 2262 2278 2294 2309 2323 2339 2355 ...
#>   ..$ : int [1:4721] 2219 2234 2248 2263 2279 2295 2310 2324 2340 2356 ...
#>   ..$ : int [1:4581] 2222 2251 2266 2282 2298 2327 2343 2359 2375 2404 ...
#>   ..$ : int [1:4463] 2229 2244 2259 2273 2290 2305 2320 2335 2350 2367 ...
#>   ..$ : int [1:5206] 2230 2245 2274 2291 2306 2321 2336 2351 2368 2382 ...
#>   ..$ : int [1:5271] 2235 2249 2264 2280 2311 2325 2341 2357 2387 2401 ...
#>   ..$ : int [1:4727] 2447 2465 2484 2505 2525 2546 2564 2583 2604 2624 ...
#>   ..$ : int [1:4601] 2449 2467 2507 2527 2548 2566 2606 2626 2647 2668 ...
#>   ..$ : int [1:4926] 2460 2480 2500 2520 2541 2559 2579 2599 2619 2640 ...
#>   ..$ : int [1:4590] 2470 2488 2530 2569 2587 2629 2671 2693 2741 2785 ...
#>   ..$ : int [1:3278] 2490 2589 2673 2696 2720 2787 2810 2833 2899 2921 ...
#>   ..$ : int [1:4979] 2649 2694 2718 2742 2764 2786 2809 2832 2855 2877 ...
#>   ..$ : int [1:3960] 4525 4539 4552 4566 4581 4596 4610 4623 4637 4652 ...
#>   ..$ : int [1:4567] 5120 5132 5143 5154 5164 5175 5187 5198 5209 5219 ...
#>   ..$ : int [1:4704] 5962 6039 6066 6092 6165 6191 6216 6289 6315 6340 ...
#>   ..$ : int [1:3572] 5968 5994 6019 6044 6072 16417 16445 16556 16583 16694 ...
#>   ..$ : int [1:4437] 5995 6020 6045 6073 6123 6147 6170 6197 6247 6271 ...
#>   ..$ : int [1:5046] 7331 7379 7404 7430 7452 7496 7518 7541 7563 7607 ...
#>   ..$ : int [1:4125] 12703 12704 12705 12706 12707 12708 12709 12710 12711 12712 ...
#>   ..$ : int [1:3779] 13088 13099 13110 13121 13132 13144 13155 13166 13177 13188 ...
#>   ..$ : int [1:4216] 13378 13413 13448 13466 13502 13538 13556 13592 13628 13646 ...
#>   ..$ : int [1:3837] 13388 13406 13440 13459 13476 13494 13530 13549 13566 13584 ...
#>   ..$ : int [1:3019] 36634 36648 36664 36681 36696 36714 36728 36743 36760 36775 ...
#>   ..$ : int [1:3830] 39920 39953 39989 40024 40058 40094 40127 40163 40198 40232 ...
#>   ..$ : int [1:3495] 39926 39959 39995 40029 40064 40100 40133 40169 40203 40238 ...
#>   ..$ : int [1:3509] 39935 39968 40004 40039 40109 40142 40178 40213 40283 40316 ...
#>   ..$ : int [1:2581] 45045 45079 45111 45145 45180 45214 45248 45280 45314 45349 ...
#>   ..$ : int [1:3401] 45552 45591 45628 45666 45704 45740 45779 45816 45855 45894 ...