Plot results of a Forward feature selection or best subset selection

A plotting function for a forward feature selection result. Each point is the mean performance of a model run. Error bars represent the standard errors from cross validation. Marked points show the best model from each number of variables until a further variable could not improve the results. If type=="selected", the contribution of the selected variables to the model performance is shown.

plot_ffs(
  ffs_model,
  plotType = "all",
  palette = rainbow,
  reverse = FALSE,
  marker = "black",
  size = 1.5,
  lwd = 0.5,
  pch = 21,
  ...
)

Arguments

ffs_model

Result of a forward feature selection see ffs

plotType

character. Either "all" or "selected"

palette

A color palette

reverse

Character. Should the palette be reversed?

marker

Character. Color to mark the best models

size

Numeric. Size of the points

lwd

Numeric. Width of the error bars

pch

Numeric. Type of point marking the best models

...

Further arguments for base plot if type="selected"

See also

Author

Marvin Ludwig and Hanna Meyer

Examples

if (FALSE) {
data(iris)
ffsmodel <- ffs(iris[,1:4],iris$Species)
plot_ffs(ffsmodel)
#plot performance of selected variables only:
plot_ffs(ffsmodel,plotType="selected")
}