Calculate Dissimilarity Index of training data

This function estimates the Dissimilarity Index (DI) of within the training data set used for a prediction model. Predictors can be weighted based on the internal variable importance of the machine learning algorithm used for model training.

trainDI(model = NA, train = NULL, variables = "all", weight = NA, folds = NULL)

Arguments

model

A train object created with caret used to extract weights from (based on variable importance) as well as cross-validation folds

train

A data.frame containing the data used for model training. Only required when no model is given

variables

character vector of predictor variables. if "all" then all variables of the model are used or if no model is given then of the train dataset.

weight

A data.frame containing weights for each variable. Only required if no model is given.

folds

Numeric or character. Folds for cross validation. E.g. Spatial cluster affiliation for each data point. Only required if no model is given.

Value

A list of class trainDI containing:

train

A data frame containing the training data

weight

A data frame with weights based on the variable importance.

variables

Names of the used variables

catvars

Which variables are categorial

scaleparam

Scaling parameters. Output from scale

trainDist_avrg

A data frame with the average eucildean distance of each training point to every other point

trainDist_avrgmean

The mean of trainDist_avrg. Used for normalizing the DI

trainDI

Dissimilarity Index of the training data

threshold

The DI threshold used for inside/outside AOA

lower_threshold

The lower DI threshold. Currently unused.

Note

This function is called within aoa to estimate the DI and AOA of new data. However, it may also be used on its own if only the DI of training data is of interest, or to facilitate a parallelization of aoa by avoiding a repeated calculation of the DI within the training data.

References

Meyer, H., Pebesma, E. (2021): Predicting into unknown space? Estimating the area of applicability of spatial prediction models. doi: 10.1111/2041-210X.13650

See also

Author

Hanna Meyer, Marvin Ludwig

Examples

if (FALSE) {
library(sf)
library(raster)
library(caret)
library(viridis)
library(latticeExtra)
library(ggplot2)

# prepare sample data:
dat <- get(load(system.file("extdata","Cookfarm.RData",package="CAST")))
dat <- aggregate(dat[,c("VW","Easting","Northing")],by=list(as.character(dat$SOURCEID)),mean)
pts <- st_as_sf(dat,coords=c("Easting","Northing"))
pts$ID <- 1:nrow(pts)
set.seed(100)
pts <- pts[1:30,]
studyArea <- stack(system.file("extdata","predictors_2012-03-25.grd",package="CAST"))[[1:8]]
trainDat <- extract(studyArea,pts,df=TRUE)
trainDat <- merge(trainDat,pts,by.x="ID",by.y="ID")

# visualize data spatially:
spplot(scale(studyArea))
plot(studyArea$DEM)
plot(pts[,1],add=TRUE,col="black")

# train a model:
set.seed(100)
variables <- c("DEM","NDRE.Sd","TWI")
model <- train(trainDat[,which(names(trainDat)%in%variables)],
trainDat$VW, method="rf", importance=TRUE, tuneLength=1,
trControl=trainControl(method="cv",number=5,savePredictions=T))
print(model) #note that this is a quite poor prediction model
prediction <- predict(studyArea,model)
plot(varImp(model,scale=FALSE))

#...then calculate the DI of the trained model:
DI = trainDI(model=model)
plot(DI)

# the DI can now be used to compute the AOA:
AOA = aoa(studyArea, model = model, trainDI = DI)
print(AOA)
plot(AOA)
}